# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CraftedBy API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/businesses": {
            "get": {
                "tags": [
                    "Businesses"
                ],
                "summary": "List all businesses",
                "operationId": "dcd44323e832c655f93f2261db5310e7",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by partial business name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "specialty",
                        "in": "query",
                        "description": "Filter by specialty",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Business"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Create a new business",
                "operationId": "90e5e72c80bec3a1734a4592f3094696",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Business"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Business created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Business"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/businesses/{id}": {
            "get": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Display a specific business",
                "operationId": "2c7e53b80a0d482985e05fe456a88a48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Business ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Business"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Business not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Update an existing business",
                "operationId": "0f3343f4b06f27e8789e49ec9a76a1a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Business ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Business"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Business updated successfully"
                    },
                    "404": {
                        "description": "Business not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Delete a business",
                "operationId": "80b021581492abfd349935bc64527461",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Business ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Business deleted successfully"
                    },
                    "404": {
                        "description": "Business not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Business": {
                "required": [
                    "name",
                    "address",
                    "postal_code",
                    "city",
                    "country",
                    "phone_number",
                    "email"
                ],
                "properties": {
                    "user_id": {
                        "description": "Owner's user ID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "Business name",
                        "type": "string"
                    },
                    "address": {
                        "description": "Business address",
                        "type": "string"
                    },
                    "postal_code": {
                        "description": "Postal code",
                        "type": "string"
                    },
                    "city": {
                        "description": "City",
                        "type": "string"
                    },
                    "country": {
                        "description": "Country",
                        "type": "string"
                    },
                    "phone_number": {
                        "description": "Phone number",
                        "type": "string"
                    },
                    "siret": {
                        "description": "SIRET number",
                        "type": "string",
                        "nullable": true
                    },
                    "craft_id": {
                        "description": "Craft ID",
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "specialty_id": {
                        "description": "Specialty ID",
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "website": {
                        "description": "Website URL",
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "description": "Email address",
                        "type": "string",
                        "format": "email"
                    },
                    "biography": {
                        "description": "Biography",
                        "type": "string",
                        "nullable": true
                    },
                    "history": {
                        "description": "History",
                        "type": "string",
                        "nullable": true
                    },
                    "theme_id": {
                        "description": "Theme ID",
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "0": "API key login",
                "type": "apiKey",
                "description": "Enter token in format \"Bearer <`token`>\" (mind the space after Bearer)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}